Updating the README and TODO
With the idea of creating a printed version in CreateSpace I can put up for sale somewhere. Not that anyone cares about this novel, of which I've failed to sell even 5 copies, but if I want to actually sell it in a bookstore I'll have to do something like that.